fix(ephemeral): CONCERNS-level fixes from the second adversarial re-review
Addresses eight WARNING/NOTE findings against the ephemeral feature after the first
adversarial-review pass (d0a8333):

- cli.py --ephemeral help text, docs/adr/0005 Scope summary, CONTEXT.md's glossary
  opening, and CHANGELOG.md's 2026-09-12 entry all still described (or, for CONTEXT.md,
  contradicted elsewhere) the retracted 3-state "0=disk/1=RAM" model. All four now state
  the shipped 2-state model plainly: EPHEMERAL=1 forces RAM (skips the fit-check, also
  enables RAM on a non-ephemeral binary); unset is auto; there is no force-disk value.
  CONTEXT.md's opening claim is now bounded ("best-effort... can fall back to disk") so it
  no longer contradicts its own Known-limitation note twelve lines down.

- test_ephemeral_disk_fallback_is_reaped's docstring overclaimed "shreds" while only
  asserting the directory disappears (a plain rm satisfies that). Narrowed the docstring to
  "reaped", and added test_disk_fallback_overwrite_flag_is_unconditional_on_staging_root: a
  source-shape test proving reapWanted/reapTarget/reapOverwrite are baked unconditionally,
  before anything downstream knows whether staging landed on RAM or the disk fallback — so
  a future change cannot silently drop --overwrite specifically on the auto-fallback path.
  Confirmed red by temporarily conditioning reapOverwrite on the staging root and observing
  the new test fail, then restored.

- The INV-CANARY-03 getEnv scanner (test_canary.py) was bypassable four ways: per-line
  scanning missed a call split across lines, an exact `getEnv(` match missed
  getenv/get_env/getENV, a single-literal regex missed a HARU literal built via string
  concatenation, and glob("*.nim") didn't recurse. Rewrote it to scan whole-file text,
  match Nim's own identifier-equality rule (first char case-sensitive, rest
  case-insensitive with underscores ignored), join every quoted token inside the call's key
  argument before the HARU check, and rglob subdirectories. Added five new tests proving
  each bypass is now caught (plus one confirming the DEV_STAGE whitelist and one confirming
  a HARU-shaped getEnv *default value* isn't a false positive). Confirmed each bypass truly
  slipped the old scanner by running its logic standalone against the same fixtures.

- cgroupAvailBytes had no fake-sysfs test for the v1 path (memory.limit_in_bytes /
  memory.usage_in_bytes), and folded "limit file absent" and "limit file present but
  unparseable" into the same -1 ("not gated") answer — so a corrupted-but-present cgroup
  file was silently treated as no limit and allowed RAM staging, asymmetric with
  shmFreeBytes/memAvailableBytes's fail-toward-disk posture. Reworked it to use fileExists
  (not the parsed value) to distinguish absent from present, so a present-but-unparseable
  limit or usage file now returns 0 ("does not fit") while a genuinely absent file or v2's
  explicit "max" still falls through as unlimited. Added a v1 fake-sysfs test and a
  parametrized (v1/v2) corrupt-file test; confirmed both red against the pre-fix logic.

- StubConfig.unpackedBytes was Nim's native `int` (32 bits on the supported armv7 target),
  read via getInt()'s narrowing int(BiggestInt) conversion — a legitimate >2^31 payload
  could truncate silently (-d:release) or fault (range checks on) before ramWouldFit's own
  overflow guard ever runs. Changed the field to int64 and parseStubConfig to
  getBiggestInt(); dropped the now-redundant int64() cast at the main.nim call site. Added a
  deterministic round-trip test (fake /proc/meminfo, since real /dev/shm capacity varies by
  host) for a value just past 2^31. Cross-compiled the real launcher for linux-armv7 with
  zig to confirm it still builds; no qemu-user emulator was available to execute it there,
  but int64 is 64 bits on every Nim target regardless, so the fix holds either way.

- Clarified the _staged_tree_bytes fallback comment in build.py: it conflated "no sidecar"
  with "corrupt sidecar" as if only the former made the launcher refuse the member; both do
  (StageError either way), the build-time estimate just can't expand either.

ruff clean; 259 tests green across test_ephemeral_safe.py, test_canary.py and
test_invariants_enforced.py, plus test_shred.py/test_ram_only.py/test_stub_config.py/
test_reap.py run separately as an extra check on the areas touched here.

+++ src/haru_pack/launcher/main.nim
@@ -152,7 +152,7 @@ proc autoEphemeralRoot(sc: StubConfig): string =
   ## Auto ephemeral (docs/adr/0005): stage to RAM only if the payload PROVABLY fits, else fall
   ## back to the persistent cache with an honest note. Never gambles RAM it cannot account for.
   when defined(linux):
-    if ramWouldFit(int64(sc.unpackedBytes)): return ramBackedRoot()
+    if ramWouldFit(sc.unpackedBytes): return ramBackedRoot()   # already int64 (stubconfig.nim)
     stderr.writeLine "haru-pack: --ephemeral: the staged payload does not fit the available " &
                      "RAM (/dev/shm + MemAvailable); staging to the persistent cache instead."
     return baseDir()
